feat: working on the UI for auth
