+++ packages/api-client/src/components/DataTable/DataTableHeader.vue
@@ -1,10 +1,12 @@
 <script setup lang="ts">
+import { tw } from '@scalar/components'
+
 import DataTableCell from './DataTableCell.vue'
 </script>
 <template>
   <DataTableCell
     is="th"
-    class="truncate font-medium px-2 py-1.5">
+    :class="tw('truncate font-medium px-2 py-1.5', $attrs.class as string)">
     <slot />
   </DataTableCell>
 </template>
