+++ packages/api-client/src/components/DataTable/DataTableRow.vue
@@ -1,5 +1,5 @@
 <template>
-  <tr class="group contents w-fit min-w-full">
+  <tr class="">
     <slot />
   </tr>
 </template>
