+++ packages/api-client/src/views/Request/RequestSection/RequestAuth/index.ts
@@ -1,1 +1,1 @@
-export { default as RequestAuth } from './RequestAuth.vue'
+export { default as RequestAuthDataTable } from './RequestAuthDataTable.vue'
